Changed timeout to use the built-in timeout of urllib2 for opening the url.
Added timeout using threading.Timer to set a timer that closes the urllib2 connection if no data is read after a certain amount of time (set in class instantiation).

Changed the default timeout to 120 seconds.

Added autoretry. It is on by default. During a download if a socket.error is caught it automatically calls self.resume to try to continue the download.